feat: added text configuration
