+++ Sources/Screens/Errors/LocalAuthErrorScreens/TouchIDErrorViewModel.swift
@@ -22,7 +22,7 @@ struct TouchIDErrorViewModel: GDSErrorViewModelV3, BaseViewModel {
         let bodyParagraph = BodyTextViewModel(text: GDSLocalisedString("app_localAuthManagerErrorBody1"))
         self.bodyContent = [
             bodyParagraph,
-            ScreenBody()
+            NumberedListView()
         ]
     }
